third_party/WebKit/LayoutTests/fast/mediastream/getusermedia-constraints.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>

// Formatting of promises is done according to
// https://www.chromium.org/blink/serviceworker/testing#TOC-Layout-Tests-Style

promise_test(function() {
  return navigator.mediaDevices.getUserMedia({})
    .then(function(s) {
        assert_unreached('getUserMedia should have failed');
      })
    .catch(function(e) {
        assert_equals(e.name, 'SyntaxError');
      });
}, 'getUserMedia() without any media types should fail');

promise_test(function() {
  return navigator.mediaDevices.getUserMedia({'audio': true})
    .then(function(s) {
         // do nothing
      })
    .catch(function(e) {
        assert_unreached('getUserMedia should have succeeded');
      });
}, 'getUserMedia() with audio=true should succeed');

promise_test(function() {
  return navigator.mediaDevices.getUserMedia(
      {audio: {'mandatory': { 'valid_but_unsupported_1': 0}}})
    .then(function(s) {
        assert_unreached('getUserMedia should have failed');
      })
    .catch(function(e) {
        assert_equals(e.name, 'ConstraintNotSatisfiedError');
      });
}, 'getUserMedia() with unsupported mandatory constraint should fail');

// The next tests document existing behavior. They seem non-conformant
// with the specs.
promise_test(function() {
  return navigator.mediaDevices.getUserMedia({video: {}})
    .then(function(s) {
        assert_unreached('getUserMedia should have failed');
      })
    .catch(function(e) {
        assert_equals(e.name, 'Error');
      });
}, 'getUserMedia() with empty constraints (neither mandatory nor optional) should fail with Error');

// The following set of tests verify behavior when trying to use the
// dictionary form of constraints. The behaviors currently expected are:
// - Unknown names in dictionary: ignored - which means Error
// - Known names but illegal syntax for value: TypeError
// - Known names and legal syntax for value: Error (no handler yet).

function check_constraints(name, constraints, expected_error) {
  promise_test(function() {
    return navigator.mediaDevices.getUserMedia({'video': constraints})
      .then(function(s) {
          assert_unreached('getUserMedia should have failed');
        })
      .catch(function(e) {
          assert_equals(e.name, expected_error);
        });
  }, name);
}

check_constraints(
  'Constraint with unsupported name gets ignored',
  {'unsupported_name': 47}, 'Error');

check_constraints(
  'Constraint with exact Long value should be parsed',
  {'height': {exact: 47}}, 'Error');
check_constraints(
  'Constraint with naked value should fail with TypeError (until supported)',
  {height: 47}, 'TypeError');

check_constraints(
  'Constraint with boolean value should be parsed',
  {'echoCancellation': {exact: true}}, 'Error');
check_constraints(
  'Constraint with boolean value should fail on naked value (until supported)',
  {'echoCancellation': true}, 'TypeError');

check_constraints(
  'Constraint with string value should work on exact with array',
  {'facingMode': {exact: ['user']}}, 'Error');
check_constraints(
  'Constraint with exact string value should fail (until supported)',
  {'facingMode': {exact: 'user'}}, 'TypeError');
check_constraints(
  'Constraint with naked string value should fail (until supported)',
  {'facingMode': 'user'}, 'TypeError');

</script>
</body>
</html>
